move region to structs
